CRUD finalizado

+++ readall.php
@@ -1,34 +1,53 @@
 <?php
 
-// Importa as configurações do aplicativo:
+/**
+ * readall.php
+ * Aplicativo que obtém todos os usuários válidos no sistema, formata e exibe 
+ * no documento.
+ */
+
+/**
+ * Importa as configurações e funções do aplicativo:
+ * Referências: https://www.w3schools.com/php/php_includes.asp
+ **/
 require('includes/config.php');
 
-// Declara / inicializa variáveis do aplicativo:
+/**
+ * Declara / inicializa variável do aplicativo:
+ * Neste caso, estamos criando um "menu principal" na variável $output que é 
+ * usada no final do código, para exibir os resultados no navegador.
+ **/
 $output = '
-<a href="/create.php">Cadastrar</a> |
-<a href="/readall.php">Listar</a>
+<a href="/readall.php">Listar</a> 
+|
+<a href="/create.php">Cadastrar</a>
 <hr>
 ';
 
-// Monta a query que obtém TODOs os usuários:
+// Monta a query que obtém TODOS os usuários válidos:
 $sql = <<<SQL
 
 SELECT uid, name, email, 
+    -- Converte a data do MySQL para o formato descrito e envia por "udatebr".
     DATE_FORMAT(udate, '%d/%m/%Y às %H:%i:%s') AS udatebr
 FROM users
+-- Não obtém usuários apagados.
 WHERE ustatus != 'deleted';
 
 SQL;
-
+ 
 // Executa a query:
 $res = $conn->query($sql);
 
-// Verifica se existem usuários:
+/**
+ * Verifica se existem usuários:
+ **/
 if ($res->num_rows == 0) :
+    // Se não exitem, exibe mensagem de erro:
     $output .= 'Oooops! Nenhum usuário cadastrado...';
 else :
 
-    // Inicia a tabela com a lista de usuários:
+    // Se existem, inicia uma tabela com a lista de usuários:
     $output .= '
     <h3>Lista de usuários</h3>
     <table border="1">
@@ -41,7 +60,10 @@ else :
     </tr>
     ';
 
-    // Loop que obtém cada usuário:
+    /**
+     * Loop que obtém cada usuário da lista do banco de dados e monta as linhas
+     * da tabela com os dados deste.
+     **/
     while ($user = $res->fetch_assoc()) :
 
         $output .= <<<HTML
@@ -54,7 +76,7 @@ else :
         <td>
             <a href="read.php?{$user['uid']}">Ver</a>
             <a href="update.php?{$user['uid']}">Editar</a>
-            <a href="delete.php?{$user['uid']}">Apagar</a>
+            <a href="delete.php?{$user['uid']}" onclick="return confirm(`Tem certeza que deseja apagar este usuário?\n\nNão tem como desfazer isso!`);">Apagar</a>
         </td>
     </tr>
